refactor(editor): extract file command controller
Move SVG/raster/netlist export, Check & Save shelf policy, and multi-file SPICE import out of App into the Editor Shell file boundary. The import report type now originates with the command that produces it instead of the Properties renderer that consumes it.

Validation: pnpm typecheck; pnpm ci:static; 25 focused export/workspace/check-save/App unit tests; focused Playwright contracts for formal SVG/PNG/PDF export, parameterized hierarchy SPICE import/re-export, and warning-reviewed NoConnect netlist export; pnpm gate:plan for owned paths; git diff --check.

Test-Impact: no-test-change — behavior is unchanged and covered by existing file-command units and focused browser import/export contracts.
